chore(ci): add feature prefix to automated upgrade PRs
This adds the `pr-title-prefix` argument to the upgrade steps in
the upgrade-bridge and upgrade-provider workflows.  In the case
where we were using the official Pulumi upgrade action, this upgrades
that action to the latest version, v0.0.12.  That version introduced
the `pr-title-prefix` argument and matches the base code of our custom
action.