chore: run make tfgen (#211)
The latest automated provider upgrade PR (#210) includes some unexpected
changes to the examples. It looks like the examples were thrown
out-of-sync by an earlier manual PR. This PR is the result of running
`make tfgen` on `main`, and aims to bring the examples back in line with
what would be produced by automation.