chore(ci): alias build_sdks as generate_sdks to fix upgrades (#281)
A couple months ago, `pulumi/upgrade-provider` was changed to expect the
existence of a `generate_sdks` task instead of a `build_sdks` task. As a
result of this change, our provider upgrade workflow is failing.

This follows [the suggestion in a comment on the `upgrade-provider`
change
PR](https://github.com/pulumi/upgrade-provider/pull/312#issuecomment-2851135278)
to alias the `build_sdks` task as `generate_sdks`, which should fix the
upgrade workflow.