feat: use same schema for Java SDK that is used for other SDKs (#318)
One cause of drift between the workflows for our Pulumi provider and the
ones for other providers is our custom Java schema patching logic. The
Pulumi-managed toolchain and CI workflows generally assume they will be
used by Pulumi; in the case of Java SDKs, this means the tools/workflows
assume that a Java SDK package will be of the form `com.pulumi.<provider
name>`. Since we are not Pulumi, we need to use `com.equinix` as the
base package instead of `com.pulumi`. The Pulumi tools provide a way to
change the base package, but the final segment of the package is always
called `<provider name>`, so if we specify a base package of
`com.equinix`, the SDK would be `com.equinix.equinix`.

To work around this, we introduced a custom, patched schema for the Java
SDK so that we generate `com.equinix.pulumi` instead of
`com.pulumi.equinix` or `com.equinix.equinix`, but such schema patching
is not a standard practice across Pulumi providers and its presence
blocks us from easily keeping our CI and toolchains aligned with the
standards. Our patching also leads to invalid examples for Java; the
examples are based on the standard, unpatched schema, so they assume the
SDK is at `com.pulumi.equinix`.

This PR removes our custom schema patching and instead sets a base
package of `com.equinix.pulumi` for the Java SDK. This results in the
SDK package changing from `com.equinix.pulumi` to
com.equinix.pulumi.equinix`; while the package name is repetitive, I
think it's worth taking that repetition in exchange for simplifying
future maintenance of this provider.